scripts/requirements.txt
oss2
python-dotenv
